Seperate editor and screen
